restructing my dish endpoints

+++ api/dish/views.py
@@ -1,7 +1,7 @@
 from flask_restx import Resource,abort
 from ..models import User,Dish
 from ..schema import dish_create,dish_namespace,dish_view,ImageSchema
-from flask_jwt_extended import jwt_required,get_jwt_identity
+from flask_jwt_extended import jwt_required,get_jwt_identity,get_jwt
 from flask_socketio import emit
 from flask import request,send_file
 from ..libs import image_helper
@@ -42,15 +42,19 @@ class Recipes(Resource):
     @jwt_required()
     def post(self):
         """creating your dish """
-        user_id=get_jwt_identity()
-        data=dish_namespace.payload
-        name=data.get("name")
-        instructions=data.get("instructions")   
-        ingredients=data.get("ingredients")
-        new_dish=Dish(name=name,instructions=instructions,ingredients=ingredients,user_id=user_id)
-        
-        new_dish.save()
-        return {"status":"success","message":"dish uploaded successfully"},201
+        user_token=get_jwt()
+        if user_token is not None:
+            user_id=get_jwt_identity()
+            data=dish_namespace.payload
+            name=data.get("name")
+            instructions=data.get("instructions")   
+            ingredients=data.get("ingredients")
+            new_dish=Dish(name=name,instructions=instructions,ingredients=ingredients,user_id=user_id)
+            
+            new_dish.save()
+            return {"status":"success","message":"dish uploaded successfully"},201
+        else:
+            return {"status":"error","message":"Jwt_token not present"},404
 
 @dish_namespace.route("/<int:id>")
 class GetUpdateDeleteDish(Resource):
@@ -78,19 +82,24 @@ class GetUpdateDeleteDish(Resource):
     @dish_namespace.response(403,"forbidden")
     def put(self,id):
         """updating dishes"""
-        user_id=get_jwt_identity()
-        if Dish.query.filter(Dish.user_id==user_id).first():
-            data=dish_namespace.payload
-            dish=Dish.query.get(id)
-            if dish:
-                dish.name=data.get("name")
-                dish.instructions=data.get("instructions")
-                dish.ingredients=data.get("ingredients")
-                dish.set_ingredients(dish.ingredients)
-                dish.update()
-                return {"status":"success","message":"dish has been updated"},200
-            abort(404,message="this post doesnt exist",status="error")
-        abort(403,message="you are not the owner of this account",status="error")
+        user_token=get_jwt()
+        if user_token is not None:
+            user_id=get_jwt_identity()
+            if Dish.query.filter(Dish.user_id==user_id).first():
+                data=dish_namespace.payload
+                dish=Dish.query.get(id)
+                if dish:
+                    dish.name=data.get("name")
+                    dish.instructions=data.get("instructions")
+                    dish.ingredients=data.get("ingredients")
+                    dish.set_ingredients(dish.ingredients)
+                    dish.update()
+                    return {"status":"success","message":"dish has been updated"},200
+                abort(404,message="this post doesnt exist",status="error")
+            abort(403,message="you are not the owner of this account",status="error")
+        else:
+            return {"status":"error","message":"Jwt_token not present"},404
+
 
     #creating an endpoint to delete dish
     @dish_namespace.response(200,"delete successful")
@@ -100,12 +109,17 @@ class GetUpdateDeleteDish(Resource):
     @jwt_required()
     def delete(self,id):
         """delete a dish"""
-        user_id=get_jwt_identity()
-        if Dish.query.filter(Dish.user_id==user_id).first():
-            dish=Dish.query.get_or_404(id)
-            dish.delete()
-            return {"status":"success","message":"dish deleted successfully"},200
-        abort(403,message="you are not the owner of this account",status="error")
+        user_token=get_jwt()
+        if user_token is not None:
+            user_id=get_jwt_identity()
+            if Dish.query.filter(Dish.user_id==user_id).first():
+                dish=Dish.query.get_or_404(id)
+                dish.delete()
+                return {"status":"success","message":"dish deleted successfully"},200
+            abort(403,message="you are not the owner of this account",status="error")
+        else:
+            return {"status":"error","message":"Jwt_token not present"},404
+
 # creating an endpoint to handle the liking and disliking a dish
 @dish_namespace.route("/likes/<int:id>")
 class postlikes(Resource):
@@ -116,23 +130,28 @@ class postlikes(Resource):
         @jwt_required()
         def post(self,id):
             """liking and disliking a dish"""
-            user_id=get_jwt_identity()
-            user=User.query.get(user_id)
-            dish=Dish.query.get_or_404(id)
-            if user is not  None:
-                if user not in dish.user_likes:
-                    dish.user_likes.append(user)
-                    dish.update()
-                    
-                else:
-                    dish.user_likes.remove(user)
-                    dish.update()
-                updated_likes=len(dish.user_likes)
-                emit('like', {'dish_id': dish.id, 'like_count': updated_likes}, broadcast=True , namespace=dish_namespace)
+            user_token=get_jwt()
+            if user_token:
+                user_id=get_jwt_identity()
+                user=User.query.get(user_id)
+                dish=Dish.query.get_or_404(id)
+                if user is not  None:
+                    if user not in dish.user_likes:
+                        dish.user_likes.append(user)
+                        dish.update()
+                        
+                    else:
+                        dish.user_likes.remove(user)
+                        dish.update()
+                    updated_likes=len(dish.user_likes)
+                    emit('like', {'dish_id': dish.id, 'like_count': updated_likes}, broadcast=True , namespace=dish_namespace)
 
-                return {"status":"success","number_of_likes":updated_likes},200
+                    return {"status":"success","number_of_likes":updated_likes},200
 
-            abort(404,message="user doesnt exist",status="error")
+                abort(404,message="user doesnt exist",status="error")
+            else:
+                return {"status":"error","message":"Jwt_token not present"},404
+
 
 @dish_namespace.route("/image/<id>")
 class Imagee(Resource):
@@ -144,29 +163,33 @@ class Imagee(Resource):
     def put(self,id):
         """uploading images"""
         # a put method used to upload the image , the image path is then saved in the database while the image is saved in a folder (static/images)
-        data=ImageSchema().load(request.files)
-        user_id=get_jwt_identity()
-        
-        avatar_path=image_helper.find_image_format(data["image"].filename)
-        if avatar_path:
+        user_token=get_jwt()
+        if user_token is not None:
+            data=ImageSchema().load(request.files)
+            user_id=get_jwt_identity()
+            
+            avatar_path=image_helper.find_image_format(data["image"].filename)
+            if avatar_path:
+                try:
+                    os.remove(avatar_path)
+                except:
+                    abort(500,message="avatar deletion failed")
             try:
-                os.remove(avatar_path)
-            except:
-                abort(500,message="avatar deletion failed")
-        try:
-            dish=Dish.query.get_or_404(id)
-            if dish.user_id!=user_id:
-                abort(403,message="you are not authorized to make this change")
-            image_path=image_helper.save_image(data["image"])
-            file_path=image_helper.get_path(image_path)
-            dish.dish_image_url=file_path
-            dish.update()
-            basename=image_helper.get_basename(image_path)
-            return {"status":"success","message":f"{basename} has been uploaded"},201
-        except UploadNotAllowed:
-            extension=image_helper.get_extension(data["image"])
-            return {"status":"error","message":f"this {extension} is not allowed"},400
+                dish=Dish.query.get_or_404(id)
+                if dish.user_id!=user_id:
+                    abort(403,message="you are not authorized to make this change")
+                image_path=image_helper.save_image(data["image"])
+                file_path=image_helper.get_path(image_path)
+                dish.dish_image_url=file_path
+                dish.update()
+                basename=image_helper.get_basename(image_path)
+                return {"status":"success","message":f"{basename} has been uploaded"},201
+            except UploadNotAllowed:
+                extension=image_helper.get_extension(data["image"])
+                return {"status":"error","message":f"this {extension} is not allowed"},400
 
+        else:
+            return {"status":"error","message":"Jwt_token not present"},404
 
 @dish_namespace.route("/images/<id>")
 class image(Resource):
@@ -203,32 +226,35 @@ class image(Resource):
     @dish_namespace.response(500,"Sever unavailable")
     @jwt_required()
     def delete(self,filename):
-        user_id=get_jwt_identity()
-        folder=f"user_{user_id}"
-        if not image_helper.is_filename_safe(filename):
-            abort(400,message=f"this {filename} is not supported",status="error")
-        try:
-            # get the current working directory
-            current_dir = os.getcwd()
+        user_token=get_jwt()
+        if user_token is not None:
+            user_id=get_jwt_identity()
+            folder=f"user_{user_id}"
+            if not image_helper.is_filename_safe(filename):
+                abort(400,message=f"this {filename} is not supported",status="error")
+            try:
+                # get the current working directory
+                current_dir = os.getcwd()
 
 
 
-            file_path=image_helper.get_path(filename,folder=folder)
-            file_path_=os.path.join(current_dir,file_path)
-            absolute_path=os.path.abspath(file_path_)
-        
-            os.remove(absolute_path)
-            return {"status":"success","message":"image deleted successfully"}
+                file_path=image_helper.get_path(filename,folder=folder)
+                file_path_=os.path.join(current_dir,file_path)
+                absolute_path=os.path.abspath(file_path_)
+            
+                os.remove(absolute_path)
+                return {"status":"success","message":"image deleted successfully"}
 
-   
+    
 
-        except FileNotFoundError:
-            abort(404,message=f"file with name {filename} not found ",status="error")
-        
-        except:
-            traceback.print_exc()
-            abort(500,message="image deletion failed",status="error")
-
+            except FileNotFoundError:
+                abort(404,message=f"file with name {filename} not found ",status="error")
+            
+            except:
+                traceback.print_exc()
+                abort(500,message="image deletion failed",status="error")
+        else:
+            return {"status":"error","message":"Jwt_token not present"},404
